Refactor shooter and turret subsystems to use new Targeting subsystem and remove old pubsub system
