--- src/main/java/frc/lib/pubsub/PubTopic.java
@@ -1,5 +1,0 @@
-package frc.lib.pubsub;
-
-public enum PubTopic {
-    SHOT_COMPENSATION
-}
